% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{biockb_store_index}
\alias{biockb_store_index}
\title{Build the Search Index of a BiocKB Store}
\usage{
biockb_store_index(store)
}
\arguments{
\item{store}{A ragnar store object.}
}
\value{
The result of \code{\link[ragnar]{ragnar_store_build_index}}
(invisibly).
}
\description{
Finalises the store by building the vector-similarity search index.
This should be called after all chunks have been inserted with
\code{\link[=biockb_store_insert]{biockb_store_insert()}}.
}
\seealso{
\code{\link[=biockb_store_create]{biockb_store_create()}}, \code{\link[=biockb_store_insert]{biockb_store_insert()}}
}
