% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocAgentRAGDB-package.R
\docType{package}
\name{BiocRAGDB-package}
\alias{BiocRAGDB}
\alias{BiocRAGDB-package}
\title{BiocRAGDB: RAG Knowledge Base for BiocAgent from Bioconductor Package Sources}
\description{
Provides a pre-built retrieval-augmented generation (RAG) database derived from the source code, vignettes, and metadata of all Bioconductor packages. The database is hosted on ExperimentHub and cached locally via BiocFileCache. It is designed to be used by the 'BiocAgent' package to ground LLM responses in real Bioconductor documentation. Includes utilities to fetch, update, and manage the local copy of the RAG database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Bioconductor/BiocRAGDB}
  \item Report bugs at \url{https://github.com/Bioconductor/BiocRAGDB/issues}
}

}
\author{
\strong{Maintainer}: Jacques Serizay \email{jacquesserizay@gmail.com}

}
\keyword{internal}
