% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{build_bioc_index}
\alias{build_bioc_index}
\title{Build the Search Index of a BiocKB Store}
\usage{
build_bioc_index(kb, ...)
}
\arguments{
\item{kb}{A \code{BiocKB} object.}
}
\value{
The \code{BiocKB} object (invisibly).
}
\description{
Finalises the store by building the vector-similarity search index.
This should be called after all chunks have been inserted with
\code{\link[=insert_bioc_kb]{insert_bioc_kb()}}.
}
\examples{
\dontrun{
kb <- BiocKB(tempfile(fileext = ".duckdb"))
# ... insert chunks ...
build_bioc_index(kb)
}

}
\seealso{
\link{BiocKB}, \code{\link[=insert_bioc_kb]{insert_bioc_kb()}}
}
