% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.find_pandoc}
\alias{.find_pandoc}
\title{Find pandoc executable, using rmarkdown's search if available}
\usage{
.find_pandoc()
}
\description{
Find pandoc executable, using rmarkdown's search if available
}
