% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.safe_markdown_chunk}
\alias{.safe_markdown_chunk}
\title{Wrapper around ragnar::markdown_chunk that retries after stripping
footnote syntax when xml2/libxml2 version mismatches cause failures.}
\usage{
.safe_markdown_chunk(md, ...)
}
\description{
Wrapper around ragnar::markdown_chunk that retries after stripping
footnote syntax when xml2/libxml2 version mismatches cause failures.
}
