% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_r_name}
\alias{extract_r_name}
\title{Best-effort name extraction from an R expression line}
\usage{
extract_r_name(expr_line, fallback_index)
}
\description{
Best-effort name extraction from an R expression line
}
